Update the Baby Names table from the SSA, add Guard
Extract the data assets to a module, and use the files in the gem as
defaults.

Users could easily switch or flip flop between datasets from
previous years. There is something very attractive about an app that
knows that a person's date of birth is from 1950, and uses the dataset
from the 1950 U.S. Social Security Administration to guess their
ethnicity, instead of the dataset from 2014. I wouldn't want to put all
those files into the gem, because it'd be huge.

Add the guard gem, and some basic plugins for it.