Refactor tests and add better coverage
Update the Code Climate test reporter, which requires running SimpleCov
locally and posting the results.

I've also added a simple .rubocop.yml, added more rubies to
the Travis CI config, and cleaned up a bunch of cruft.

The README now has to-dos, which is pretty much a list of things I don't
like about this code.